refactor #2
