+++ Cyan/parser.py
@@ -1,135 +1,9 @@
-from Cyan.tokens import T
-from Cyan.utils import InvalidSyntaxError
-
-# for type hinting
 from typing import Optional
-from Cyan.tokens import Token
-from Cyan.utils import Pos
 
-
-class Node:
-    pos_start: Optional[Pos]
-    pos_end: Optional[Pos]
-
-    def set_pos(self, pos_start: Pos, pos_end: Optional[Pos] = None):
-        self.pos_start = pos_start
-        if pos_end is not None:
-            self.pos_end = pos_end
-
-        return self
-
-    def __repr__(self):
-        return type(self).__name__
-
-
-class StatementsNode(Node):
-    def __init__(self, statements, pos_start, pos_end):
-        self.statements = statements
-        super().set_pos(pos_start, pos_end)
-
-    def __repr__(self):
-        stmts = ", ".join(repr(statement) for statement in self.statements)
-        return f"Statements({stmts})"
-
-
-class NumberNode(Node):
-    def __init__(self, tok):
-        self.tok = tok
-        super().set_pos(tok.pos_start, tok.pos_end)
-
-    def __repr__(self):
-        return str(self.tok)
-
-
-class LiteralNode(Node):
-    def __init__(self, tok):
-        self.tok = tok
-        super().set_pos(tok.pos_start, tok.pos_end)
-
-    def __repr__(self):
-        return self.tok.value
-
-
-class StringNode(Node):
-    def __init__(self, tok):
-        self.tok = tok
-        super().set_pos(tok.pos_start, tok.pos_end)
-
-
-class BinOpNode(Node):
-    def __init__(self, left, oper, right):
-        self.left = left
-        self.oper = oper
-        self.right = right
-        super().set_pos(left.pos_start, right.pos_end)
-
-    def __repr__(self):
-        return f"({self.left}, {self.oper}, {self.right})"
-
-
-class UnaryOpNode(Node):
-    def __init__(self, oper, node):
-        self.oper = oper
-        self.node = node
-        super().set_pos(oper.pos_start, node.pos_end)
-
-    def __repr__(self):
-        return f"({self.oper}, {self.node})"
-
-
-class VarAccessNode(Node):
-    def __init__(self, var_name):
-        self.var_name = var_name
-        super().set_pos(var_name.pos_start, var_name.pos_end)
-
-    def __repr__(self):
-        return f"({self.var_name})"
-
-
-class VarAssignNode(Node):
-    def __init__(self, var_name, value):
-        self.var_name = var_name
-        self.value = value
-        super().set_pos(var_name.pos_start, value.pos_end)
-
-    def __repr__(self):
-        return f"({self.var_name} = {self.value})"
-
-
-class IfBlockNode(Node):
-    def __init__(self, case: tuple, else_expr):
-        self.case = case
-        self.else_expr = else_expr
-        super().set_pos(case[0].pos_start)
-
-    def __repr__(self):
-        return f"(if {self.case[0]} then {self.case[1]} else {self.else_expr})"
-
-
-class WhileNode(Node):
-    def __init__(self, condition, body):
-        self.condition = condition
-        self.body = body
-        super().set_pos(condition.pos_start, body.pos_end)
-
-    def __repr__(self):
-        return f"(while {self.condition} do {self.body})"
-
-
-class FuncDefNode(Node):
-    def __init__(self, name: str, parameters: list[Token], body: Node):
-        self.name = name or "[lambda]"
-        self.parameters = parameters
-        self.body = body
-
-
-class FuncCallNode(Node):
-    def __init__(self, node_to_call: Node, arguments: list[Node]):
-        self.node_to_call = node_to_call
-        self.arguments = arguments
-
-    def __repr__(self):
-        return f"(FuncCall:{self.node_to_call})"
+from cyan.tokens import T
+from cyan.exceptions import InvalidSyntaxError
+import cyan.ast as ast
+from cyan.tokens import Token
 
 
 class ParseResult:
@@ -148,7 +22,7 @@ class ParseResult:
         self.advancements += 1
         pass
 
-    def register(self, res) -> Node:
+    def register(self, res) -> ast.Node:
         self.advancements += res.advancements
         if res.error:
             self.error = res.error
@@ -199,7 +73,7 @@ class Parser:
     def parse(self):
         res = self.statements()
         if res.error is None and (not self.cur_tok.is_type(T.EOF, T.NEWLINE)):
-            from Cyan.utils import Printer
+            from cyan.utils import Printer
 
             Printer.debug_p(self.cur_tok)
             res.failure(
@@ -245,7 +119,9 @@ class Parser:
                 continue
             statements.append(statement)
 
-        return res.success(StatementsNode(statements, pos_start, self.cur_tok.pos_end))
+        return res.success(
+            ast.StatementsNode(statements, pos_start, self.cur_tok.pos_end)
+        )
 
     def expr(self):
         res = ParseResult()
@@ -280,7 +156,9 @@ class Parser:
             if res.error:
                 return res
 
-            return res.success(VarAssignNode(var_name, expr))
+            return res.success(
+                ast.VarAssignNode(var_name, expr)
+            )
 
         node = res.register(
             self.bin_oper(self.comp_expr, ((T.KW, "and"), (T.KW, "or")))
@@ -348,7 +226,7 @@ class Parser:
             self.advance()
 
             return res.success(
-                FuncCallNode(atom, args).set_pos(atom.pos_start, pos_end)
+                ast.FuncCallNode(atom, args).set_pos(atom.pos_start, pos_end)
             )
         return res.success(atom)
 
@@ -359,12 +237,12 @@ class Parser:
         if tok.is_type(T.INT, T.FLOAT):
             res.register_adv()
             self.advance()
-            return res.success(NumberNode(tok))
+            return res.success(ast.NumberNode(tok))
 
         elif tok.is_type(T.LITERAL):
             res.register_adv()
             self.advance()
-            return res.success(LiteralNode(tok))
+            return res.success(ast.LiteralNode(tok))
 
         elif tok.is_type(T.L_PAREN):
             res.register_adv()
@@ -386,12 +264,12 @@ class Parser:
         elif tok.is_type(T.IDENTIFIER):
             res.register_adv()
             self.advance()
-            return res.success(VarAccessNode(tok))
+            return res.success(ast.VarAccessNode(tok))
 
         elif tok.is_type(T.STRING):
             res.register_adv()
             self.advance()
-            return res.success(StringNode(tok))
+            return res.success(ast.StringNode(tok))
 
         elif tok.is_equals(T.KW, "if"):
             node = res.register(self.if_expr())
@@ -426,7 +304,7 @@ class Parser:
             factor = res.register(self.factor())
             if res.error:
                 return res
-            return res.success(UnaryOpNode(tok, factor))
+            return res.success(ast.UnaryOpNode(tok, factor))
 
         return self.power()
 
@@ -450,7 +328,10 @@ class Parser:
             if res.error:
                 return res
 
-            return res.success(UnaryOpNode(op_tok, node))
+            return res.success(
+                ast.UnaryOpNode(op_tok, node)
+            )
+            
         node = res.register(
             self.bin_oper(self.arith_expr, (T.EE, T.NE, T.LT, T.GT, T.LTE, T.GTE))
         )
@@ -481,7 +362,7 @@ class Parser:
             if res.error:
                 return res
 
-            left = BinOpNode(left, op_tok, right)
+            left = ast.BinOpNode(left, op_tok, right)
 
         return res.success(left)
 
@@ -522,7 +403,7 @@ class Parser:
         if res.error:
             return res
 
-        return res.success(IfBlockNode((cond, expr), else_expr))
+        return res.success(ast.IfBlockNode((cond, expr), else_expr))
 
     def func_def(self):
         # self.cur_tok is KW:fun
@@ -608,7 +489,7 @@ class Parser:
         self.advance()
 
         return res.success(
-            FuncDefNode(name, parameters, statements).set_pos(
+            ast.FuncDefNode(name, parameters, statements).set_pos(
                 pos_start, statements.pos_end
             )
         )
@@ -648,10 +529,12 @@ class Parser:
         res.register_adv()
         self.advance()
 
-        return res.success(WhileNode(cond, statements).set_pos(cond.pos_start, p_end))
+        return res.success(
+            ast.WhileNode(cond, statements).set_pos(cond.pos_start, p_end)
+        )
 
 
-def make_ast(tokens):
+def parse_ast(tokens):
     parser = Parser(tokens)
     res = parser.parse()
     return res.node, res.error
